// src/gameData/enum.tsx
export enum NPCNames {
  OldLady = "Old Lady",
  Bartender = "Bartender",
}

export enum QuestStages {
  NotStarted = "Not Started",
  InProgress = "In Progress",
  Completed = "Completed",
}

export enum QuestNames {
  SecretIngredient = "Secret Ingredient",
}

export enum Locations {
  Tavern = "Tavern",
  outsideTavern = "Ironwood Avenue",
  BackGate = "Ironwood Gate",
  EmeraldPlaza = "Emerald Plaza",
  CityGates = "Arcanthria City Gates",
  EvershadeForest = "Evershade Forest",
  TheCrystalCaves = "Crystal Caves",
  ElderglowLake = "Elderglow Lake",
  Eldervale = "Eldervale",
  CircleOfTheGods = "Circle Of The Gods",
  AetherPeak = "Aether Peak",
  EaglesPeak = "Eagles Peak",
  ArmorShop = "Armor Shop",
  tavernRoom = "Tavern Room",
  WhisperingWoods = "Whispering Woods",
  RiverbendHarbor = "Riverbend Harbor",
  harbor = "Harbor",
  BlackVoid = "Black Void",
  PotionShop = "Potion Shop",
}

export enum Loot {
  FeatherLeafSkeleton = "Feather Leaf Skeleton",
  FeatherLeafFluid = "Feather Leaf Fluid",
  SirenSkull = "Siren Skull",
  SirenScale = "Siren Scale",
}